+++ src/bin/email_api_dir/admin_ops/diag_handlers/mod.rs
@@ -1,0 +1,9 @@
+#![allow(unused_imports, dead_code)]
+
+pub mod deliverability;
+pub mod observability;
+pub mod security;
+
+pub use deliverability::*;
+pub use observability::*;
+pub use security::*;
